Create linguistic
The benefit of this class is that despite of the language also iconography and wordings in the same language is made dynamic. The gained benefit is a GUI which impacts in a sophisticated manner the users emotions. If only language is required then use better another class which is better tested.